+++ src/fireworks/training/cookbook/utils/__init__.py
@@ -1,6 +1,7 @@
 """Cookbook utilities -- infrastructure, losses, data, logging, and more.
 
-RL-specific utilities (GRPO/DAPO/GSPO losses, TIS, R3, batching) live in
+RL-specific utilities (losses, training loop, PP recommendation,
+importance sampling, router replay) live in
 ``fireworks.training.cookbook.utils.rl``.
 """
 
@@ -16,6 +17,7 @@ __all__ = [
     "StepCallback",
     "WandBConfig",
     "compute_advantages",
+    "compute_pass_at_k",
     "create_trainer_job",
     "resolve_and_apply_shape",
     "encode_text",
@@ -25,14 +27,19 @@ __all__ = [
     "load_preference_dataset",
     "log_metrics_json",
     "make_dpo_loss_fn",
+    "make_orpo_loss_fn",
     "make_batch_sft_loss_fn",
     "make_sft_loss_fn",
     "setup_deployment",
     "setup_resume",
     "setup_training_client",
     "setup_wandb",
+    "flush_timing",
+    "timed",
+    "timer",
     "validate_config",
     "validate_preflight",
+    "validate_streaming_config",
     "wandb_finish",
     "wandb_log",
 ]
@@ -51,6 +58,7 @@ from fireworks.training.cookbook.utils.infra import (
     setup_training_client,
     resolve_and_apply_shape,
 )
+from fireworks.training.cookbook.utils.timer import timed, timer, flush_timing
 from fireworks.training.cookbook.utils.client import ReconnectableClient
 from fireworks.training.cookbook.utils.config import (
     DEFAULT_ADAM,
@@ -66,6 +74,7 @@ from fireworks.training.cookbook.utils.config import (
 from fireworks.training.cookbook.utils.losses import (
     make_dpo_loss_fn,
     make_sft_loss_fn,
+    make_orpo_loss_fn,
     make_batch_sft_loss_fn,
 )
 from fireworks.training.cookbook.utils.resume import setup_resume
@@ -74,5 +83,6 @@ from fireworks.training.cookbook.utils.logging import (
     setup_wandb,
     wandb_finish,
     log_metrics_json,
+    compute_pass_at_k,
 )
-from fireworks.training.cookbook.utils.validation import validate_config, validate_preflight
+from fireworks.training.cookbook.utils.validation import validate_config, validate_preflight, validate_streaming_config